+++ pulsar/apps/data/stores/__init__.py
@@ -1,2 +1,3 @@
 from . import redis
 from .pulsards import *
+from .mongodb import *